Rename app files
